+++ app/controllers/web/2fa/index.js
@@ -1,0 +1,3 @@
+const recovery = require('./recovery');
+
+module.exports = { recovery };
