+++ app/views/2fa/otp-recovery.pug
@@ -10,7 +10,7 @@ block body
             .text-center
               h1.card-title.h4= t('Two-Factor Authentication Recovery')
               p= t('Please enter a two-factor recovery key to login.')
-            form(action=`${ctx.locale}/2fa/recovery-keys`, method="POST", autocomplete="off").ajax-form
+            form(action=`${ctx.locale}/2fa/recovery/keys`, method="POST", autocomplete="off").ajax-form
               input(type="hidden", name="_csrf", value=ctx.csrf)
               .form-group.floating-label
                 input#input-text.form-control.form-control-lg(type="text", autocomplete="off", name="recovery_passcode", placeholder=t(''))
