+++ app/views/admin/users/retrieve.pug
@@ -26,6 +26,10 @@ block body
                   option(value='user', selected=result.group === 'user') User
                   option(value='admin', selected=result.group === 'admin') Admin
                 label(for='input-group')= t('Group')
+              if result[config.userFields.twoFactorEnabled]
+                .form-check
+                  input#two-factor-enabled(type='checkbox', name=config.userFields.twoFactorEnabled, value='true' checked)
+                  label(for='input-two-factor-enabled')= t('Two-Factor Enabled')
             .card-footer.text-right
               button(type='reset').btn.btn-secondary= t('Reset')
               button(type='submit').btn.btn-primary= t('Save')
