+++ routes/web/2fa.js
@@ -10,10 +10,10 @@ router.use(policies.ensureLoggedIn);
 router
   .get('/otp/login', render('2fa/otp-login'))
   .post('/otp/login', web.auth.loginOtp)
-  .post('/recovery', web.auth.recovery)
+  .post('/recovery', web.twoFactor.recovery.recover)
   .get('/recovery/verify', render('2fa/verify'))
-  .post('/recovery/verify', web.auth.recoveryVerify)
-  .get('/recovery-keys', render('2fa/otp-recovery'))
-  .post('/recovery-keys', web.auth.recoveryKey);
+  .post('/recovery/verify', web.twoFactor.recovery.verify)
+  .get('/recovery/keys', render('2fa/otp-recovery'))
+  .post('/recovery/keys', web.auth.recoveryKey);
 
 module.exports = router;
